+++ frontend/packages/client/src/pages/ProposalCreate.js
@@ -10,11 +10,11 @@ import {
   PropCreateStepThree,
   PropCreateStepTwo,
 } from 'components/ProposalCreate';
-import { useProposal } from 'hooks';
+import { useProposalCreateMutation } from 'hooks';
 import { isStartTimeValid, parseDateToServer } from 'utils';
 
 export default function ProposalCreatePage() {
-  const { createProposal, data, loading, error } = useProposal();
+  const { createProposal, data, loading, error } = useProposalCreateMutation();
   const [modalError, setModalError] = useState(null);
   const {
     user: { addr: creatorAddr },
